fix(color picker): store separate RGB and HSV values
Add a private HSV reactive attribute to the `ColorPicker` to store the
current HSV value.

The problem with only storing the RGB value is that it is lossy. For
example, when the color is black (0, 0, 0) and you change the hue,
the color won't update which causes the widgets to get out of sync.